git-svn-id: svn://srv-dc/svn2/Citrus@171565 20bf7d6a-bd21-9d4d-a3c4-a2c14bb8c68a

+++ Lime/Source/Application/Application.cs
@@ -106,7 +106,7 @@ namespace Lime
 		public virtual void OnCreate() {}
 		public virtual void OnGLCreate() {}
 		public virtual void OnGLDestroy() {}
-		public virtual void OnUpdateFrame(double delta) {}
+		public virtual void OnUpdateFrame(int delta) {}
 		public virtual void OnRenderFrame() {}
 		public virtual void OnDeviceRotated() {}
 		public virtual DeviceOrientation GetSupportedDeviceOrientations() { return DeviceOrientation.LandscapeLeft; }
